CAPTCHA+OPTIMIZATION+ICON

+++ src/views/mesdonnees.vue
@@ -1,22 +1,38 @@
 <script setup>
 import { useRouter } from 'vue-router';  // Import useRouter to access the router instance
 import Button from 'primevue/button';
+import Toast from 'primevue/toast';
+import { useToast } from 'primevue/usetoast';
 
 const router = useRouter();  // Initialize the router
+const toast = useToast();
 
 function effacerStockageLocal(){
     if (localStorage.length == 0){
-        alert("Aucune donnée n'est stockée sur votre navigateur");
+        showInfo()
         return router.go(-1);  // Go back to the previous page
     }
 
+    const confirmation = confirm("Êtes-vous sûr de vouloir supprimer toutes vos données ?");
+    if (!confirmation) {
+        return;  // If the user cancels, do nothing
+    }
+
     localStorage.clear();
-    alert("Vos données ont été supprimées avec succès");
+    showSuccess();
     return router.go(-1);  // Go back to the previous page
 }
+    const showSuccess = () => {
+        toast.add({ severity: 'success', summary: 'Données supprimées!', detail: 'Vos données ont été supprimé avec succès.', life: 2500 });
+    };
+
+    const showInfo = () => {
+        toast.add({ severity: 'info', summary: 'Erreur', detail: 'Vous n\'avez aucune donnée à supprimer.', life: 2500 });
+    };
 </script>
 
 <template>
+    <Toast></Toast>
     <div id="dataPageContent">
         <div id="title_container">
             <h1 id="title">Vos données</h1>
@@ -33,11 +49,12 @@ function effacerStockageLocal(){
 </template>
 
 <style scoped>
-    #dataPageContent {
+    div#dataPageContent {
         display: flex;
         flex-direction: column;
         align-items: center;
         justify-content: center;
+        margin-top:50px;
     }
 
     #disclaimer {
@@ -47,6 +64,16 @@ function effacerStockageLocal(){
     }
 
     #suppButtonContainer {
-        margin-top: 100px;
+        margin-top: 10vh;
+    }
+
+    #suppButton {
+        width: 200px;
+        height:5vh;
+        font-family: inherit;
+    }
+
+    #suppButton:hover {
+        border-color:var(--red-800)
     }
 </style>
